feat(config): add duplicated keys validation, create sh script to launch config generation
