+++ src/Utils/ConfigGeneration/Constants.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+namespace Fawaz\Utils\ConfigGeneration;
+
+class Constants {
+    static $pathToAssets = "./runtime-data/media/assets/"; 
+    static $pathForEditing = "./src/config/backend-config-for-editing/";
+    static $configUrlBase = "https://media.getpeer.eu/assets/";
+
+
+    static $inputFileNameSuffix = "-editable";
+    static $extension = ".json";
+}
