+++ src/config/backend-config-for-editing/response-codes-editable.json
@@ -1,5 +1,5 @@
 {
-    "10601": {
+    "10401": {
         "comment": "Success: Contact created successfully",
         "userFriendlyComment": "Your message has been sent successfully!"
     },
